--- src/components/Main/ImgComponents/Second/Second.css
@@ -1,13 +1,0 @@
-.image {
-  position: relative;
-  height: 100vh;
-}
-
-.content {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: absolute;
-  height: 100vh;
-  width: 100vh;
-}
