--- src/components/Main/ImgComponents/Second/Second.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { Parallax } from 'react-parallax';
-import Tools from '../../../../assets/images/two.jpg';
-
-const Second = () => (
-  <Parallax bgImage={Tools} strength={300}>
-    <div style={{ height: '400px' }} />
-  </Parallax>
-);
-
-export default Second;
