+++ src/components/Main/Main.jsx
@@ -1,20 +1,23 @@
 import React from 'react';
-import First from './ImgComponents/First/First';
-import Second from './ImgComponents/Second/Second';
-import Third from './ImgComponents/Third/Third';
+import Hero from '../Hero/Hero';
 import FirstVisit from '../FirstVisit/FirstVisit';
 import OurServices from '../OurServices/OurServices';
 import Places from '../Places/Places';
 import ContactForm from '../ContactForm/ContactForm';
+import ParallaxImage from '../ParallaxImage/ParallaxImage';
+import parallaxImageOne from '../../assets/images/parallax/two.jpg';
+import parallaxImageTwo from '../../assets/images/parallax/three.jpg';
+import Navbar from '../Navbar/Navbar';
 
 const Main = () => {
   return (
     <div>
-      <First />
+      <Navbar />
+      <Hero />
       <OurServices />
-      <Second />
+      <ParallaxImage image={parallaxImageOne} />
       <FirstVisit />
-      <Third />
+      <ParallaxImage image={parallaxImageTwo} />
       <ContactForm />
       <Places />
     </div>
