+++ src/components/ParallaxImage/ParallaxImage.css
@@ -1,0 +1,3 @@
+.parallax-content {
+  height: 400px;
+}
