+++ src/components/ScrollToTop/ScrollToTop.css
@@ -6,7 +6,7 @@
   width: 50px;
 }
 
-.button {
+.scroll-to-top-button {
   border-radius: 50%;
   height: 50px;
   width: 50px;
