+++ src/components/TelephoneButton/TelephoneButton.css
@@ -1,0 +1,11 @@
+.telephone-button {
+  position: fixed;
+  bottom: 50%;
+  right: 0px;
+  height: 40px;
+  width: 110px;
+  background-color: black;
+  color: white;
+  cursor: pointer;
+  border: none;
+}
