+++ src/components/TelephoneButton/TelephoneButton.jsx
@@ -1,47 +1,21 @@
 import React from 'react';
-import { useEffect, useState } from 'react';
+import './TelephoneButton.css';
+
+import { elementId } from '../../constants/elementId';
 
 const TelephoneButton = () => {
-  const [BackToTop, setBackToTop] = useState(false);
-  useEffect(() => {
-    const handleScroll = () => {
-      if (window.scrollY > 100) {
-        setBackToTop(true);
-      } else {
-        setBackToTop(false);
-      }
-    };
-    window.addEventListener('scroll', handleScroll);
-    return () => window.removeEventListener('scroll', handleScroll); // Clean up listener
-  }, []);
-
   const scrollToSection = () => {
-    const section = document.getElementById('umów się na wizytę');
+    const section = document.getElementById(elementId.contact);
+
     if (section) {
       section.scrollIntoView({ behavior: 'smooth' });
     }
   };
+
   return (
-    <div>
-      {BackToTop && (
-        <button
-          style={{
-            position: 'fixed',
-            bottom: '50%',
-            right: '0px',
-            height: '40px',
-            width: '110px',
-            backgroundColor: 'black',
-            color: 'white',
-            cursor: 'pointer',
-            border: 'none',
-          }}
-          onClick={scrollToSection}
-        >
-          Kontakt
-        </button>
-      )}
-    </div>
+    <button className='telephone-button' onClick={scrollToSection}>
+      Kontakt
+    </button>
   );
 };
